feat(aci): Move project and environment to detector step 2
Removes project and environments from step 1 and makes them available on step 2.

This helps with editing detectors since previously you could not edit project or environment.

Preserves selected project id from the detector list as the default for a new detector.
