+++ static/app/views/detectors/list.tsx
@@ -9,6 +9,7 @@ import SentryDocumentTitle from 'sentry/components/sentryDocumentTitle';
 import {ActionsProvider} from 'sentry/components/workflowEngine/layout/actions';
 import ListLayout from 'sentry/components/workflowEngine/layout/list';
 import {useWorkflowEngineFeatureGate} from 'sentry/components/workflowEngine/useWorkflowEngineFeatureGate';
+import {ALL_ACCESS_PROJECTS} from 'sentry/constants/pageFilters';
 import {IconAdd} from 'sentry/icons';
 import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
@@ -100,10 +101,21 @@ function TableHeader() {
 
 function Actions() {
   const organization = useOrganization();
+  const {selection} = usePageFilters();
+
+  let project: number | undefined;
+  if (selection.projects) {
+    // Find the first selected project id that is not the all access project
+    project = selection.projects.find(pid => pid !== ALL_ACCESS_PROJECTS);
+  }
+
   return (
     <Fragment>
       <LinkButton
-        to={`${makeMonitorBasePathname(organization.slug)}new/`}
+        to={{
+          pathname: `${makeMonitorBasePathname(organization.slug)}new/`,
+          query: project ? {project} : undefined,
+        }}
         priority="primary"
         icon={<IconAdd />}
         size="sm"
