+++ static/app/views/detectors/new.tsx
@@ -15,6 +15,7 @@ import {useWorkflowEngineFeatureGate} from 'sentry/components/workflowEngine/use
 import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
 import type {DetectorType} from 'sentry/types/workflowEngine/detectors';
+import {useLocation} from 'sentry/utils/useLocation';
 import {useNavigate} from 'sentry/utils/useNavigate';
 import useOrganization from 'sentry/utils/useOrganization';
 import useProjects from 'sentry/utils/useProjects';
@@ -23,7 +24,6 @@ import {makeMonitorBasePathname} from 'sentry/views/detectors/pathnames';
 
 interface NewDetectorFormData {
   detectorType: DetectorType;
-  environment: string;
   project: string;
 }
 
@@ -31,8 +31,11 @@ export default function DetectorNew() {
   const navigate = useNavigate();
   const organization = useOrganization();
   useWorkflowEngineFeatureGate({redirect: true});
+  const location = useLocation();
   const {projects} = useProjects();
 
+  const projectIdFromLocation =
+    typeof location.query.project === 'string' ? location.query.project : undefined;
   const defaultProject = projects.find(p => p.isMember) ?? projects[0];
 
   const newMonitorName = t('New Monitor');
@@ -46,7 +49,6 @@ export default function DetectorNew() {
           query: {
             detectorType: data.detectorType,
             project: data.project,
-            environment: data.environment,
           },
         });
       }}
@@ -54,8 +56,7 @@ export default function DetectorNew() {
       initialData={
         {
           detectorType: 'metric_issue',
-          project: defaultProject?.id ?? '',
-          environment: '',
+          project: projectIdFromLocation ?? defaultProject?.id ?? '',
         } satisfies NewDetectorFormData
       }
     >
